Database.PostgreSQL.Simple.Arrays

Description

A Postgres array parser and pretty-printer.

Documentation

arrayFormat :: Char -> Parser ArrayFormat

Parse one of three primitive field formats: array, quoted and plain.

data ArrayFormat

Constructors

Array [ArrayFormat]

Plain ByteString

Quoted ByteString

array :: Char -> Parser [ArrayFormat]

quoted :: Parser ByteString

Recognizes a quoted string.

plain :: Char -> Parser ByteString

Recognizes a plain string literal, not containing quotes or brackets and not containing the delimiter character.

fmt :: Char -> ArrayFormat -> ByteString

Format an array format item, using the delimiter character if the item is itself an array.

delimit :: Char -> [ArrayFormat] -> ByteString

Format a list of array format items, inserting the appropriate delimiter between them. When the items are arrays, they will be delimited with commas; otherwise, they are delimited with the passed-in-delimiter.

fmt' :: Bool -> Char -> ArrayFormat -> ByteString

Format an array format item, using the delimiter character if the item is itself an array, optionally applying quoting rules. Creates copies for safety when used in FromField instances.

esc :: ByteString -> ByteString

Escape a string according to Postgres double-quoted string format.