Database.PostgreSQL.Simple.HStore

Description

Parsers and printers for hstore, a extended type bundled with PostgreSQL providing finite maps from text strings to text strings. See https://www.postgresql.org/docs/9.5/static/hstore.html for more information.

Note that in order to use this type, a database superuser must install it by running a sql script in the share directory. This can be done on PostgreSQL 9.1 and later with the command CREATE EXTENSION hstore. See https://www.postgresql.org/docs/9.5/static/contrib.html for more information.

Documentation

newtype HStoreList

Constructors

HStoreList

Fields

fromHStoreList :: [(Text, Text)]

newtype HStoreMap

Constructors

HStoreMap

Fields

fromHStoreMap :: Map Text Text

class ToHStore a where

Methods

toHStore :: a -> HStoreBuilder

data HStoreBuilder

Represents valid hstore syntax.

toBuilder :: HStoreBuilder -> Builder

toLazyByteString :: HStoreBuilder -> ByteString

hstore :: (ToHStoreText a, ToHStoreText b) => a -> b -> HStoreBuilder

parseHStoreList :: ByteString -> Either String HStoreList

class ToHStoreText a where

Methods

toHStoreText :: a -> HStoreText

data HStoreText

Represents escape text, ready to be the key or value to a hstore value