Database.PostgreSQL.Simple.LargeObjects

Description

Support for PostgreSQL's Large Objects; see https://www.postgresql.org/docs/9.5/static/largeobjects.html for more information.

Note that Large Object File Descriptors are only valid within a single database transaction, so if you are interested in using anything beyond loCreat, loCreate, and loUnlink, you will need to run the entire sequence of functions in a transaction. As loImport and loExport are simply C functions that call loCreat, loOpen, loRead, and loWrite, and do not perform any transaction handling themselves, they also need to be wrapped in an explicit transaction.

Documentation

loCreat :: Connection -> IO Oid

loCreate :: Connection -> Oid -> IO Oid

loImport :: Connection -> FilePath -> IO Oid

loImportWithOid :: Connection -> FilePath -> Oid -> IO Oid

loExport :: Connection -> Oid -> FilePath -> IO ()

loOpen :: Connection -> Oid -> IOMode -> IO LoFd

loWrite :: Connection -> LoFd -> ByteString -> IO Int

loRead :: Connection -> LoFd -> Int -> IO ByteString

loSeek :: Connection -> LoFd -> SeekMode -> Int -> IO Int

loTell :: Connection -> LoFd -> IO Int

loTruncate :: Connection -> LoFd -> Int -> IO ()

loClose :: Connection -> LoFd -> IO ()

loUnlink :: Connection -> Oid -> IO ()

newtype Oid

Constructors

Oid CUInt

data LoFd

LoFd is a Large Object (pseudo) File Descriptor. It is understood by libpq but not by operating system calls.

data IOMode

See openFile

Constructors

ReadMode

WriteMode

AppendMode

ReadWriteMode

data SeekMode

A mode that determines the effect of hSeek hdl mode i.

Constructors

AbsoluteSeek
the position of hdl is set to i.

RelativeSeek
the position of hdl is set to offset i from the current position.

SeekFromEnd
the position of hdl is set to offset i from the end of the file.