Database.PostgreSQL.Simple.Range

Description

Documentation

data RangeBound a

Represents boundary of a range

Constructors

NegInfinity

Inclusive !a

Exclusive !a

PosInfinity

data PGRange a

Generic range type

Constructors

PGRange !(RangeBound a) !(RangeBound a)

empty :: PGRange a

isEmpty :: Ord a => PGRange a -> Bool

Is a range empty? If this returns True, then the contains predicate will always return False. However, if this returns False, it is not necessarily true that there exists a point for which contains returns True. Consider PGRange (Excludes 2) (Excludes 3) :: PGRange Int, for example.

isEmptyBy :: (a -> a -> Ordering) -> PGRange a -> Bool

contains :: Ord a => PGRange a -> a -> Bool

Does a range contain a given point? Note that in some cases, this may not correspond exactly with a server-side computation. Consider UTCTime for example, which has a resolution of a picosecond, whereas postgresql's timestamptz types have a resolution of a microsecond. Putting such Haskell values into the database will result in them being rounded, which can change the value of the containment predicate.

containsBy :: (a -> a -> Ordering) -> PGRange a -> a -> Bool

fromFieldRange :: Typeable a => FieldParser a -> FieldParser (PGRange a)