Database.PostgreSQL.Simple.TypeInfo.Macro

Description

A Template Haskell macro for efficiently checking membership in a set of type oids.

Documentation

mkCompats :: [TypeInfo] -> ExpQ

Returns an expression that has type Oid -> Bool, true if the oid is equal to any one of the typoids of the given TypeInfos.

inlineTypoid :: TypeInfo -> ExpQ

Literally substitute the typoid of a TypeInfo expression. Returns an expression of type Oid. Useful because GHC tends not to fold constants.