Database.PostgreSQL.Simple.TypeInfo.Static

Description

This module contains portions of the pg_type table that are relevant to postgresql-simple and are believed to not change between PostgreSQL versions.

Documentation

data TypeInfo

A structure representing some of the metadata regarding a PostgreSQL type, mostly taken from the pg_type table.

Constructors

Basic

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

Array

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

typelem :: !TypeInfo

Range

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

rngsubtype :: !TypeInfo

Composite

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

typrelid :: !Oid

attributes :: !(Vector Attribute)

staticTypeInfo :: Oid -> Maybe TypeInfo

bool :: TypeInfo

boolOid :: Oid

bytea :: TypeInfo

byteaOid :: Oid

char :: TypeInfo

charOid :: Oid

name :: TypeInfo

nameOid :: Oid

int8 :: TypeInfo

int8Oid :: Oid

int2 :: TypeInfo

int2Oid :: Oid

int4 :: TypeInfo

int4Oid :: Oid

regproc :: TypeInfo

regprocOid :: Oid

text :: TypeInfo

textOid :: Oid

oid :: TypeInfo

oidOid :: Oid

tid :: TypeInfo

tidOid :: Oid

xid :: TypeInfo

xidOid :: Oid

cid :: TypeInfo

cidOid :: Oid

xml :: TypeInfo

xmlOid :: Oid

point :: TypeInfo

pointOid :: Oid

lseg :: TypeInfo

lsegOid :: Oid

path :: TypeInfo

pathOid :: Oid

box :: TypeInfo

boxOid :: Oid

polygon :: TypeInfo

polygonOid :: Oid

line :: TypeInfo

lineOid :: Oid

cidr :: TypeInfo

cidrOid :: Oid

float4 :: TypeInfo

float4Oid :: Oid

float8 :: TypeInfo

float8Oid :: Oid

unknown :: TypeInfo

unknownOid :: Oid

circle :: TypeInfo

circleOid :: Oid

money :: TypeInfo

moneyOid :: Oid

macaddr :: TypeInfo

macaddrOid :: Oid

inet :: TypeInfo

inetOid :: Oid

bpchar :: TypeInfo

bpcharOid :: Oid

varchar :: TypeInfo

varcharOid :: Oid

date :: TypeInfo

dateOid :: Oid

time :: TypeInfo

timeOid :: Oid

timestamp :: TypeInfo

timestampOid :: Oid

timestamptz :: TypeInfo

timestamptzOid :: Oid

interval :: TypeInfo

intervalOid :: Oid

timetz :: TypeInfo

timetzOid :: Oid

bit :: TypeInfo

bitOid :: Oid

varbit :: TypeInfo

varbitOid :: Oid

numeric :: TypeInfo

numericOid :: Oid

refcursor :: TypeInfo

refcursorOid :: Oid

record :: TypeInfo

recordOid :: Oid

void :: TypeInfo

voidOid :: Oid

array_record :: TypeInfo

array_recordOid :: Oid

regprocedure :: TypeInfo

regprocedureOid :: Oid

regoper :: TypeInfo

regoperOid :: Oid

regoperator :: TypeInfo

regoperatorOid :: Oid

regclass :: TypeInfo

regclassOid :: Oid

regtype :: TypeInfo

regtypeOid :: Oid

uuid :: TypeInfo

uuidOid :: Oid

json :: TypeInfo

jsonOid :: Oid

jsonb :: TypeInfo

jsonbOid :: Oid

int2vector :: TypeInfo

int2vectorOid :: Oid

oidvector :: TypeInfo

oidvectorOid :: Oid

array_xml :: TypeInfo

array_xmlOid :: Oid

array_json :: TypeInfo

array_jsonOid :: Oid

array_line :: TypeInfo

array_lineOid :: Oid

array_cidr :: TypeInfo

array_cidrOid :: Oid

array_circle :: TypeInfo

array_circleOid :: Oid

array_money :: TypeInfo

array_moneyOid :: Oid

array_bool :: TypeInfo

array_boolOid :: Oid

array_bytea :: TypeInfo

array_byteaOid :: Oid

array_char :: TypeInfo

array_charOid :: Oid

array_name :: TypeInfo

array_nameOid :: Oid

array_int2 :: TypeInfo

array_int2Oid :: Oid

array_int2vector :: TypeInfo

array_int2vectorOid :: Oid

array_int4 :: TypeInfo

array_int4Oid :: Oid

array_regproc :: TypeInfo

array_regprocOid :: Oid

array_text :: TypeInfo

array_textOid :: Oid

array_tid :: TypeInfo

array_tidOid :: Oid

array_xid :: TypeInfo

array_xidOid :: Oid

array_cid :: TypeInfo

array_cidOid :: Oid

array_oidvector :: TypeInfo

array_oidvectorOid :: Oid

array_bpchar :: TypeInfo

array_bpcharOid :: Oid

array_varchar :: TypeInfo

array_varcharOid :: Oid

array_int8 :: TypeInfo

array_int8Oid :: Oid

array_point :: TypeInfo

array_pointOid :: Oid

array_lseg :: TypeInfo

array_lsegOid :: Oid

array_path :: TypeInfo

array_pathOid :: Oid

array_box :: TypeInfo

array_boxOid :: Oid

array_float4 :: TypeInfo

array_float4Oid :: Oid

array_float8 :: TypeInfo

array_float8Oid :: Oid

array_polygon :: TypeInfo

array_polygonOid :: Oid

array_oid :: TypeInfo

array_oidOid :: Oid

array_macaddr :: TypeInfo

array_macaddrOid :: Oid

array_inet :: TypeInfo

array_inetOid :: Oid

array_timestamp :: TypeInfo

array_timestampOid :: Oid

array_date :: TypeInfo

array_dateOid :: Oid

array_time :: TypeInfo

array_timeOid :: Oid

array_timestamptz :: TypeInfo

array_timestamptzOid :: Oid

array_interval :: TypeInfo

array_intervalOid :: Oid

array_numeric :: TypeInfo

array_numericOid :: Oid

array_timetz :: TypeInfo

array_timetzOid :: Oid

array_bit :: TypeInfo

array_bitOid :: Oid

array_varbit :: TypeInfo

array_varbitOid :: Oid

array_refcursor :: TypeInfo

array_refcursorOid :: Oid

array_regprocedure :: TypeInfo

array_regprocedureOid :: Oid

array_regoper :: TypeInfo

array_regoperOid :: Oid

array_regoperator :: TypeInfo

array_regoperatorOid :: Oid

array_regclass :: TypeInfo

array_regclassOid :: Oid

array_regtype :: TypeInfo

array_regtypeOid :: Oid

array_uuid :: TypeInfo

array_uuidOid :: Oid

array_jsonb :: TypeInfo

array_jsonbOid :: Oid

int4range :: TypeInfo

int4rangeOid :: Oid

_int4range :: TypeInfo

_int4rangeOid :: Oid

numrange :: TypeInfo

numrangeOid :: Oid

_numrange :: TypeInfo

_numrangeOid :: Oid

tsrange :: TypeInfo

tsrangeOid :: Oid

_tsrange :: TypeInfo

_tsrangeOid :: Oid

tstzrange :: TypeInfo

tstzrangeOid :: Oid

_tstzrange :: TypeInfo

_tstzrangeOid :: Oid

daterange :: TypeInfo

daterangeOid :: Oid

_daterange :: TypeInfo

_daterangeOid :: Oid

int8range :: TypeInfo

int8rangeOid :: Oid

_int8range :: TypeInfo

_int8rangeOid :: Oid