Database.PostgreSQL.Simple.Vector

Description

query variants returning Vector.

Documentation

query :: (ToRow q, FromRow r) => Connection -> Query -> q -> IO (Vector r)

Perform a SELECT or other SQL query that is expected to return results. All results are retrieved and converted before this function returns.

query_ :: FromRow r => Connection -> Query -> IO (Vector r)

A version of query that does not perform query substitution.

queryWith :: ToRow q => RowParser r -> Connection -> Query -> q -> IO (Vector r)

A version of query taking parser as argument

queryWith_ :: RowParser r -> Connection -> Query -> IO (Vector r)

A version of query_ taking parser as argument

returning :: (ToRow q, FromRow r) => Connection -> Query -> [q] -> IO (Vector r)

Execute INSERT ... RETURNING, UPDATE ... RETURNING, or other SQL query that accepts multi-row input and is expected to return results.

returningWith :: ToRow q => RowParser r -> Connection -> Query -> [q] -> IO (Vector r)

A version of returning taking parser as argument