Database.PostgreSQL.Simple.Cursor

Description

Types

data Cursor

Cursor within a transaction.

Cursor management

declareCursor :: Connection -> Query -> IO Cursor

Declare a temporary cursor. The cursor is given a unique name for the given connection.

closeCursor :: Cursor -> IO ()

Close the given cursor.

Folding over rows from a cursor

foldForward :: FromRow r => Cursor -> Int -> (a -> r -> IO a) -> a -> IO (Either a a)

Fold over a chunk of rows, calling the supplied fold-like function on each row as it is received. In case the cursor is exhausted, a Left value is returned, otherwise a Right value is returned.

foldForwardWithParser :: Cursor -> RowParser r -> Int -> (a -> r -> IO a) -> a -> IO (Either a a)

Fold over a chunk of rows from the given cursor, calling the supplied fold-like function on each row as it is received. In case the cursor is exhausted, a Left value is returned, otherwise a Right value is returned.