Database.PostgreSQL.Simple.ToRow

Description

The ToRow typeclass, for rendering a collection of parameters to a SQL query.

Predefined instances are provided for tuples containing up to ten elements.

Documentation

class ToRow a where

A collection type that can be turned into a list of rendering Actions.

Instances should use the toField method of the ToField class to perform conversion of each element of the collection.

Methods

toRow :: a -> [Action]

toRow :: (Generic a, GToRow (Rep a)) => a -> [Action]