Database.PostgreSQL.Simple.TypeInfo.Static

Description

This module contains portions of the pg_type table that are relevant to postgresql-simple and are believed to not change between PostgreSQL versions.

Documentation

data TypeInfo

A structure representing some of the metadata regarding a PostgreSQL type, mostly taken from the pg_type table.

Constructors

Basic

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

Array

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

typelem :: !TypeInfo

Range

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

rngsubtype :: !TypeInfo

Composite

Fields

typoid :: !Oid

typcategory :: !Char

typdelim :: !Char

typname :: !ByteString

typrelid :: !Oid

attributes :: !(Vector Attribute)

staticTypeInfo :: Oid -> Maybe TypeInfo

bool :: TypeInfo

bytea :: TypeInfo

char :: TypeInfo

name :: TypeInfo

int8 :: TypeInfo

int2 :: TypeInfo

int4 :: TypeInfo

regproc :: TypeInfo

text :: TypeInfo

oid :: TypeInfo

tid :: TypeInfo

xid :: TypeInfo

cid :: TypeInfo

xml :: TypeInfo

point :: TypeInfo

lseg :: TypeInfo

path :: TypeInfo

box :: TypeInfo

polygon :: TypeInfo

line :: TypeInfo

cidr :: TypeInfo

float4 :: TypeInfo

float8 :: TypeInfo

unknown :: TypeInfo

circle :: TypeInfo

money :: TypeInfo

macaddr :: TypeInfo

inet :: TypeInfo

bpchar :: TypeInfo

varchar :: TypeInfo

date :: TypeInfo

time :: TypeInfo

timestamp :: TypeInfo

timestamptz :: TypeInfo

interval :: TypeInfo

timetz :: TypeInfo

bit :: TypeInfo

varbit :: TypeInfo

numeric :: TypeInfo

refcursor :: TypeInfo

record :: TypeInfo

void :: TypeInfo

array_record :: TypeInfo

regprocedure :: TypeInfo

regoper :: TypeInfo

regoperator :: TypeInfo

regclass :: TypeInfo

regtype :: TypeInfo

uuid :: TypeInfo

json :: TypeInfo

jsonb :: TypeInfo

int2vector :: TypeInfo

oidvector :: TypeInfo

array_xml :: TypeInfo

array_json :: TypeInfo

array_line :: TypeInfo

array_cidr :: TypeInfo

array_circle :: TypeInfo

array_money :: TypeInfo

array_bool :: TypeInfo

array_bytea :: TypeInfo

array_char :: TypeInfo

array_name :: TypeInfo

array_int2 :: TypeInfo

array_int2vector :: TypeInfo

array_int4 :: TypeInfo

array_regproc :: TypeInfo

array_text :: TypeInfo

array_tid :: TypeInfo

array_xid :: TypeInfo

array_cid :: TypeInfo

array_oidvector :: TypeInfo

array_bpchar :: TypeInfo

array_varchar :: TypeInfo

array_int8 :: TypeInfo

array_point :: TypeInfo

array_lseg :: TypeInfo

array_path :: TypeInfo

array_box :: TypeInfo

array_float4 :: TypeInfo

array_float8 :: TypeInfo

array_polygon :: TypeInfo

array_oid :: TypeInfo

array_macaddr :: TypeInfo

array_inet :: TypeInfo

array_timestamp :: TypeInfo

array_date :: TypeInfo

array_time :: TypeInfo

array_timestamptz :: TypeInfo

array_interval :: TypeInfo

array_numeric :: TypeInfo

array_timetz :: TypeInfo

array_bit :: TypeInfo

array_varbit :: TypeInfo

array_refcursor :: TypeInfo

array_regprocedure :: TypeInfo

array_regoper :: TypeInfo

array_regoperator :: TypeInfo

array_regclass :: TypeInfo

array_regtype :: TypeInfo

array_uuid :: TypeInfo

array_jsonb :: TypeInfo

int4range :: TypeInfo

_int4range :: TypeInfo

numrange :: TypeInfo

_numrange :: TypeInfo

tsrange :: TypeInfo

_tsrange :: TypeInfo

tstzrange :: TypeInfo

_tstzrange :: TypeInfo

daterange :: TypeInfo

_daterange :: TypeInfo

int8range :: TypeInfo

_int8range :: TypeInfo